requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
defusedxml==0.7.1
exceptiongroup==1.1.2
Flask==2.3.2
Flask-Compress==1.13
fonttools==4.41.1
frozenlist==1.4.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jira==3.5.2
jtutils==0.0.8
keyring==24.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
more-itertools==10.0.0
multidict==6.0.4
nose==1.3.7
numpy==1.25.1
oauthlib==3.2.2
outcome==1.2.0
packaging==23.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
sortedcontainers==2.4.0
soupsieve==2.4.1
tenacity==8.2.2
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
Werkzeug==2.3.6
wsproto==1.2.0
xlrd==2.0.1
xmltodict==0.13.0
yarl==1.9.2
zipp==3.16.2